+++ CHANGELOG.md
@@ -37,7 +37,7 @@
 
 ### Fixes
 
-* `[jest-runner]` Assign `process.env.JEST_WORKER_ID="0"` when in runInBand mode
+* `[jest-runner]` Assign `process.env.JEST_WORKER_ID="1"` when in runInBand mode
   ([#5860](https://github.com/facebook/jest/pull/5860))
 * `[jest-cli]` Add descriptive error message when trying to use
   `globalSetup`/`globalTeardown` file that doesn't export a function.